Move webrtc to its own hook
